Migrate conversation_files to new format.
